refactor: Coin관련 기능 UserCoinService로 분리
